+++ breeze-http/src/main/kotlin/HttpConfig.kt
@@ -11,7 +11,6 @@ import javax.net.ssl.*
 
 /**
  * Http的配置。
- * @see HttpClient.Builder
  */
 data class HttpConfig(
 	val cookieHandler: CookieHandler? = null,
@@ -26,20 +25,4 @@ data class HttpConfig(
 	val authenticator: Authenticator? = null,
 	val encoding: String = "UTF-8",
 ) {
-	class Builder : com.windea.breezeframework.core.Builder<HttpConfig> {
-		var cookieHandler: CookieHandler? = null
-		var connectTimeout: Duration? = null
-		var sslContext: SSLContext? = null
-		var sslParameters: SSLParameters? = null
-		var executor: Executor? = null
-		var followRedirects: HttpClient.Redirect? = null
-		var version: HttpClient.Version? = null
-		var priority: Int? = null
-		var proxy: ProxySelector? = null
-		var authenticator: Authenticator? = null
-		var encoding: String = "UTF-8"
-
-		override fun build() = HttpConfig(cookieHandler, connectTimeout, sslContext, sslParameters, executor,
-			followRedirects, version, priority, proxy, authenticator, encoding)
-	}
 }