+++ breeze-serialization/src/main/kotlin/extension/SerializationInternalExtensions.kt
@@ -48,39 +48,27 @@ internal fun <T> Array<T>.joinToAsync(buffer:Appendable,separator:String, transf
 }
 
 
-private const val kotlinxJsonClassName = "kotlinx.serialization.json.Json"
-private const val jacksonJsonClassName = "com.fasterxml.jackson.databind.json.JsonMapper"
-private const val gsonClassName = "com.google.gson.Gson"
-private const val fastjsonClassName = "com.alibaba.fastjson.JSON"
-
 internal val defaultJsonSerializer: JsonSerializer = when {
-	presentInClassPath(jacksonJsonClassName) -> JacksonJsonSerializer()
-	presentInClassPath(kotlinxJsonClassName) -> KotlinxJsonSerializer()
-	presentInClassPath(gsonClassName) -> GsonSerializer()
-	presentInClassPath(fastjsonClassName) -> FastJsonSerializer()
+	presentInClassPath("com.fasterxml.jackson.databind.json.JsonMapper") -> JacksonJsonSerializer()
+	presentInClassPath("kotlinx.serialization.json.Json") -> KotlinxJsonSerializer()
+	presentInClassPath("com.google.gson.Gson") -> GsonSerializer()
+	presentInClassPath("com.alibaba.fastjson.JSON") -> FastJsonSerializer()
 	else -> BreezeJsonSerializer()
 }
 
-private const val jacksonYamlClassName = "com.fasterxml.jackson.dataformat.yaml.YAMLMapper"
-private const val snackYamlClassName = "org.yaml.snakeyaml.Yaml"
-
 internal val defaultYamlSerializer: YamlSerializer = when {
-	presentInClassPath(jacksonYamlClassName) -> JacksonYamlSerializer()
-	presentInClassPath(snackYamlClassName) -> SnakeYamlSerializer()
+	presentInClassPath("com.fasterxml.jackson.dataformat.yaml.YAMLMapper") -> JacksonYamlSerializer()
+	presentInClassPath("org.yaml.snakeyaml.Yaml") -> SnakeYamlSerializer()
 	else -> BreezeYamlSerializer()
 }
 
-private const val jacksonXmlClassName = "com.fasterxml.jackson.dataformat.xml.XmlMapper"
-
 internal val defaultXmlSerializer: XmlSerializer = when {
-	presentInClassPath(jacksonXmlClassName) -> JacksonXmlSerializer()
+	presentInClassPath("com.fasterxml.jackson.dataformat.xml.XmlMapper") -> JacksonXmlSerializer()
 	else -> BreezeXmlSerializer()
 }
 
-private const val jacksonPropertiesClassName = "com.fasterxml.jackson.dataformat.javaprop.JavaPropsMapper"
-
 internal val defaultPropertiesSerializer: PropertiesSerializer = when {
-	presentInClassPath(jacksonPropertiesClassName) -> JacksonPropertiesSerializer()
+	presentInClassPath("com.fasterxml.jackson.dataformat.javaprop.JavaPropsMapper") -> JacksonPropertiesSerializer()
 	else -> BreezePropertiesSerializer()
 }