+++ breeze-serialization/src/main/kotlin/io/DataReader.kt
@@ -1,0 +1,21 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.serialization.io
+
+import java.lang.reflect.*
+
+/**
+ * 数据的读取器。
+ */
+internal interface DataReader{
+	/**
+	 * 读取数据。
+	 */
+	fun <T> read(value:String,type:Class<T>):T
+
+	/**
+	 * 读取数据。
+	 */
+	fun <T> read(value:String,type: Type):T
+}