+++ breeze-serialization/src/main/kotlin/serializer/DataSerializer.kt
@@ -13,6 +13,11 @@ import java.lang.reflect.*
  * 其具体实现可能需要依赖第三方库，如`gson`，`fastjson`，`jackson`和`kotlinx-serialization`。
  *
  * @see DataFormat
+ * @see JsonSerializer
+ * @see YamlSerializer
+ * @see XmlSerializer
+ * @see PropertiesSerializer
+ * @see CsvSerializer
  */
 interface DataSerializer:Serializer<String> {
 	/**