+++ breeze-serialization/src/main/kotlin/serializer/DelegateSerializer.kt
@@ -6,5 +6,10 @@ package com.windea.breezeframework.serialization.serializer
 /**
  * 由第三方库委托实现的序列化器。
  * 其具体实现需要依赖第三方库，如`gson`，`fastjson`，`jackson`和`kotlinx-serialization`。
+ *
+ * @see JacksonSerializer
+ * @see KotlinxSerializer
+ * @see GsonSerializer
+ * @see FastJsonSerializer
  */
 interface DelegateSerializer