+++ breeze-serialization/src/main/kotlin/serializer/JacksonSerializer.kt
@@ -6,6 +6,10 @@ package com.windea.breezeframework.serialization.serializer
 /**
  * 由Jackson委托实现的序列化器。
  *
- * @see com.fasterxml.jackson.databind.ObjectMapper
+ * @see com.fasterxml.jackson.databind.json.JsonMapper
+ * @see JacksonJsonSerializer
+ * @see JacksonYamlSerializer
+ * @see JacksonXmlSerializer
+ * @see JacksonPropertiesSerializer
  */
 interface JacksonSerializer : DataSerializer,DelegateSerializer