+++ breeze-serialization/src/main/kotlin/serializer/Serializer.kt
@@ -10,6 +10,9 @@ import java.lang.reflect.*
  * 序列化器。
  *
  * 序列化器用于基于特定的格式，对数据进行序列化和反序列化。
+ *
+ * @see DataSerializer
+ * @see MapLikeSerializer
  */
 interface Serializer<V>{
 	/**